requirements.txt
django
django-tinymce
fontawesomefree
django-crispy-forms
django-recaptcha
six